// src/api/models/insightLogResDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { InsightLogResDTOCategory } from './insightLogResDTOCategory';

export interface InsightLogResDTO {
  category: InsightLogResDTOCategory;
  createdAt: string;
  id: number;
  title: string;
  description: string;
  activityNames: string[];
}